# Fast instruction/base pointers