# Represent instruction and base pointers as VM attributes

Instruction and base pointers were always the `bp` and `ip` attributes of the last element of `callFrames` (defined as `vector<CallFrame>`). This probably forces the compiler to store and retrieve them via memory access with high cache miss rate (i.e. `bp` and `ip` will be in different address each time a new call stack in pushed/popped).


This change will represent them as members within the VM itself, and only store and load them into the call stack on a `CALL` or `RETURN` opcode. This enables better cache caching behaviour and can even encourage the compiler to store them in registers if it's smart enough.

### 32nd fibonacci number benchmark

**`master`**
```
bin/lisp tests/recur.scm  0.61s user 0.00s system 99% cpu 0.617 total
```

**This PR**
```
bin/lisp tests/recur.scm  0.45s user 0.00s system 99% cpu 0.452 total
```

26.7% performance improvements